Separate gpio implementations between some families
